New strategy: use original types in annotations
Code to generate type definitions for generated monomorphic types (_1,
_2, etc.) removed, since it didn't work properly. Instead, I'm keeping
track of the original types when the monomorphic ones are created and
their names are standardized, so that when Atdgen sees _1, it knows to
annotate with the type that _1 came from. This required a change to the
Atd repository as well. But it works! Hooray!