Support for rtd generation
The switch that I promised before. This is a more integrated documentation
solution. One that we will not have to re-generate and upload manually. We still
need to re-organize the documentation for things to make more sense, but this is
a start. To view the docs locally, you will need to install sphinx and the rtd
theme (read the docs)